requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# LangChain + LangGraph
langchain==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
langgraph==0.2.55
langgraph-checkpoint-postgres==1.0.9  # Changed from 0.1.9
langsmith==0.1.147

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
aiosqlite==0.20.0  # for test SQLite

# Redis
upstash-redis==1.1.0

# Qdrant + Embeddings
qdrant-client==1.11.3
fastembed==0.3.6

# GitHub --- DIRECT SDK
PyGithub==2.5.0

# Web Search --- DIRECT SDK
tavily-python==0.5.0

# E2B Sandbox
e2b-code-interpreter==1.0.3

# Safety
guardrails-ai>=0.6.0
# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
tiktoken==0.8.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
